Rename to global_body_planner, add skeleton for local_footstep_planner
