+++ global_body_planner/config/simple_terrain.yaml
