+++ global_body_planner/include/global_body_planner/functions.h
