+++ global_body_planner/launch/example_with_terrain.launch
@@ -3,19 +3,19 @@
         <param name="robot_description" command="cat $(find spirit_visualizer)/urdf/spirit40_gaz_RML.urdf" />
         
         <!-- Launch the nodes to publish the base tf and joint positions -->
-        <node name="body_planner" pkg="body_planner" type="body_planner_node" output="screen"/>
+        <node name="global_body_planner" pkg="global_body_planner" type="global_body_planner_node" output="screen"/>
 
 
 		<!-- Launch the grid map simple demo node -->
-		<node pkg="body_planner" type="simple_terrain" name="simple_terrain" output="screen" />
+		<node pkg="global_body_planner" type="simple_terrain" name="simple_terrain" output="screen" />
 		<!-- Launch the grid map visualizer -->
  		<!-- <node pkg="grid_map_demos" type="simple_demo" name="grid_map_simple_demo" output="screen" /> -->
 
 		<node pkg="grid_map_visualization" type="grid_map_visualization" name="grid_map_visualization" output="screen">
-		  <rosparam command="load" file="$(find body_planner)/config/simple_terrain.yaml" />
+		  <rosparam command="load" file="$(find global_body_planner)/config/simple_terrain.yaml" />
 		</node>
 
         <!-- Launch the nodes to publish all the robot tfs and visualize in rviz -->
-        <node name="rviz" pkg="rviz" type="rviz" args="-d $(find body_planner)/rviz/example_with_terrain.rviz"/>
+        <node name="rviz" pkg="rviz" type="rviz" args="-d $(find global_body_planner)/rviz/example_with_terrain.rviz"/>
 
 </launch>
