+++ global_body_planner/package.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>spirit_body_planner</name>
+  <name>global_body_planner</name>
   <version>0.0.0</version>
-  <description>The spirit_body_planner package</description>
+  <description>The global_body_planner package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="joe@todo.todo">joe</maintainer>
+  <maintainer email="jnorby@andrew.cmu.edu">Joe</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
