+++ global_body_planner/rviz/example.rviz
