+++ global_body_planner/rviz/example_with_terrain.rviz
