+++ global_body_planner/src/planner_class.cpp
@@ -1,4 +1,4 @@
-#include "body_planner/planner_class.h"
+#include "global_body_planner/planner_class.h"
 #include <queue>
 #include <chrono>
