+++ global_body_planner/src/simple_terrain_node.cpp
