+++ local_footstep_planner/src/local_footstep_planner.cpp
@@ -1,0 +1,13 @@
+#include "local_footstep_planner.h"
+
+LocalFootstepPlanner::LocalFootstepPlanner(ros::NodeHandle nh) {
+  _nh = nh;
+}
+
+void LocalFootstepPlanner::loop() {
+while (ros::ok())
+{
+  ros::spinOnce();
+}
+
+}
